+++ Data/Models/OrderDetail.cs
@@ -2,10 +2,10 @@
 {
     public class OrderDetail
     {
-        public int id { get; set; }
-        public int orderId { get; set; }
-        public int laptopId { get; set; }
-        public uint price { get; set; }
+        public long id { get; set; }
+        public long orderId { get; set; }
+        public long laptopId { get; set; }
+        public ulong price { get; set; }
         public virtual Laptop laptop { get; set; }
         public virtual Order order { get; set; }
     }
